+++ gui_engine/engine/acc_sw_rle_cover.c
@@ -24,10 +24,10 @@
 void rle_cover_blit_2_rgb565(draw_img_t *image, struct gui_dispdev *dc,
                              struct gui_rect *rect)
 {
-    int16_t x_start = 0;
-    int16_t x_end = 0;
-    int16_t y_start = 0;
-    int16_t y_end = 0;
+    int32_t x_start = 0;
+    int32_t x_end = 0;
+    int32_t y_start = 0;
+    int32_t y_end = 0;
     int16_t source_w = image->img_w;
 
     if (draw_img_target_area(image, dc, rect, &x_start, &x_end, &y_start, &y_end) == false)
