+++ gui_engine/engine/draw_img.h
@@ -65,7 +65,7 @@ uint32_t draw_img_get_pixel_byte(draw_img_t *img, IMG_SOURCE_MODE_TYPE src_mode)
 void draw_img_cache(draw_img_t *img, IMG_SOURCE_MODE_TYPE src_mode);
 void draw_img_free(draw_img_t *img, IMG_SOURCE_MODE_TYPE src_mode);
 bool draw_img_target_area(draw_img_t *image, struct gui_dispdev *dc, gui_rect_t *rect,
-                          int16_t *x_start, int16_t *x_end, int16_t *y_start, int16_t *y_end);
+                          int32_t *x_start, int32_t *x_end, int32_t *y_start, int32_t *y_end);
 extern void (* draw_img_acc_prepare_cb)(struct draw_img *image, gui_rect_t *rect);
 extern void (* draw_img_acc_end_cb)(struct draw_img *image);
